+++ packages/extension/src/flowFileUtils.ts
@@ -1,8 +1,11 @@
 import * as vscode from 'vscode';
 
 const FLOW_DEFINITION_SUFFIX = '.flow.json';
+const SUBFLOW_DEFINITION_SUFFIX = '-subflow.json';
+const WORKFLOW_DEFINITION_SUFFIX = '-workflow.json';
 const DIAGRAM_SUFFIX = '.diagram.json';
 const WORKFLOWS_SEGMENT = 'workflows';
+const WORKFLOWS_CAPITAL_SEGMENT = 'Workflows';
 const JSON_SUFFIX = '.json';
 
 function toNormalizedPath(uri: vscode.Uri): string {
@@ -26,18 +29,11 @@ export function isWorkflowJsonUri(uri: vscode.Uri): boolean {
   const hasJsonSuffix = normalizedPath.endsWith(JSON_SUFFIX);
   const isDiagram = isDiagramFile(uri);
   const hasWorkflowsSegment = segments.includes(WORKFLOWS_SEGMENT);
+  const hasWorkflowsCapitalSegment = segments.includes(WORKFLOWS_CAPITAL_SEGMENT);
   const isFlowDefinition = normalizedPath.endsWith(FLOW_DEFINITION_SUFFIX);
+  const isSubflowDefinition = normalizedPath.endsWith(SUBFLOW_DEFINITION_SUFFIX);
+  const isWorkflowDefinition = normalizedPath.endsWith(WORKFLOW_DEFINITION_SUFFIX);
 
-  console.log('isWorkflowJsonUri check:', {
-    path: uri.path,
-    normalizedPath,
-    segments,
-    hasJsonSuffix,
-    isDiagram,
-    hasWorkflowsSegment,
-    isFlowDefinition,
-    result: hasJsonSuffix && !isDiagram && (isFlowDefinition || hasWorkflowsSegment)
-  });
 
   if (!hasJsonSuffix) {
     return false;
@@ -47,15 +43,21 @@ export function isWorkflowJsonUri(uri: vscode.Uri): boolean {
     return false;
   }
 
-  // Accept all JSON files that are either:
+  // Accept JSON files that are either:
   // 1. .flow.json files, OR
-  // 2. Any .json file in a workflows directory, OR
-  // 3. Any .json file (for broader compatibility)
-  return isFlowDefinition || hasWorkflowsSegment || true; // Allow all JSON files
+  // 2. -subflow.json files, OR
+  // 3. -workflow.json files, OR
+  // 4. Any .json file in a workflows directory (lowercase), OR
+  // 5. Any .json file in a Workflows directory (capital W)
+  return isFlowDefinition || isSubflowDefinition || isWorkflowDefinition || hasWorkflowsSegment || hasWorkflowsCapitalSegment;
 }
 
 export function isFlowDefinitionUri(uri: vscode.Uri): boolean {
-  return toNormalizedPath(uri).endsWith(FLOW_DEFINITION_SUFFIX) || isWorkflowJsonUri(uri);
+  const normalizedPath = toNormalizedPath(uri);
+  return normalizedPath.endsWith(FLOW_DEFINITION_SUFFIX) ||
+         normalizedPath.endsWith(SUBFLOW_DEFINITION_SUFFIX) ||
+         normalizedPath.endsWith(WORKFLOW_DEFINITION_SUFFIX) ||
+         isWorkflowJsonUri(uri);
 }
 
 export function getDiagramUri(flowUri: vscode.Uri): vscode.Uri {
@@ -65,6 +67,14 @@ export function getDiagramUri(flowUri: vscode.Uri): vscode.Uri {
     return flowUri.with({ path: path.replace(/\.flow\.json$/i, DIAGRAM_SUFFIX) });
   }
 
+  if (path.toLowerCase().endsWith(SUBFLOW_DEFINITION_SUFFIX)) {
+    return flowUri.with({ path: path.replace(/-subflow\.json$/i, DIAGRAM_SUFFIX) });
+  }
+
+  if (path.toLowerCase().endsWith(WORKFLOW_DEFINITION_SUFFIX)) {
+    return flowUri.with({ path: path.replace(/-workflow\.json$/i, DIAGRAM_SUFFIX) });
+  }
+
   if (path.toLowerCase().endsWith(JSON_SUFFIX)) {
     return flowUri.with({ path: path.replace(/\.json$/i, DIAGRAM_SUFFIX) });
   }
@@ -72,11 +82,20 @@ export function getDiagramUri(flowUri: vscode.Uri): vscode.Uri {
   return flowUri.with({ path: `${path}${DIAGRAM_SUFFIX}` });
 }
 
-export const FLOW_FILE_GLOBS = ['**/*.flow.json', '**/workflows/**/*.json', '**/*.json'];
+export const FLOW_FILE_GLOBS = [
+  '**/*.flow.json',
+  '**/*-subflow.json',
+  '**/*-workflow.json',
+  '**/workflows/**/*.json',
+  '**/Workflows/**/*.json'
+];
 export const FLOW_AND_DIAGRAM_GLOBS = [
   '**/*.flow.json',
+  '**/*-subflow.json',
+  '**/*-workflow.json',
   '**/*.diagram.json',
   '**/workflows/**/*.json',
   '**/workflows/**/*.diagram.json',
-  '**/*.json'
+  '**/Workflows/**/*.json',
+  '**/Workflows/**/*.diagram.json'
 ];
